install.sh: Add deprecation warning
[why]
I guess people do not know this probably, or gloss over it when reading the readme.md.

We would need to throw together something new that maybe utilizes
bin/scripts/fetch-archives.sh

But then, cloning the repo is no good idea in any situation. And how would
people get the new script without cloning. Not sure it's worth it.

Lets see if people complain with this warning.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>